Display preloaded data directly from SSR
The trick is to get the data from the cache that have been preloaded.
This makes Svelte skip the loading part in SSR.

Currently I'm refetching the json data on the client side. This would be
better if I used some kind of window.__SVITE_DATA__ constant. One less
roundtrip.